High score implemented. Fixed play button bug
Fixed bug where if you clicked play it wouldn't give you the stats of the player. Also made it so that high scores are a thing now and are saved to the database. Added golden enemy (just need to spawn it in) that will give you 100% supermeter on destruction.

+++ Assets/Scripts/Player/SupermeterBar.cs
@@ -9,7 +9,7 @@ public class SupermeterBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetMaxSuper(int super)
+    public void SetMaxSuper(float super)
     {
         slider.maxValue = super;
         //slider.value = super;
@@ -17,7 +17,7 @@ public class SupermeterBar : MonoBehaviour
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetSuper(int super)
+    public void SetSuper(float super)
     {
         slider.value = super;
